Splitting out CharStream.
